requirements.txt
cohere>=5.0.0
pypdf>=4.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.0.0
pytest>=8.0.0
